+++ plugin/kubernetes/main.go
@@ -28,23 +28,8 @@ func main() {
 			if err != nil {
 				return nil, err
 			}
-			return &libplugin.SshPiperPluginConfig{
-				NextAuthMethodsCallback: func(_ libplugin.ConnMetadata) ([]string, error) {
-					return plugin.supportedMethods()
-				},
-
-				PasswordCallback: func(conn libplugin.ConnMetadata, password []byte) (*libplugin.Upstream, error) {
-					return plugin.findAndCreateUpstream(conn, string(password), nil)
-				},
-
-				PublicKeyCallback: func(conn libplugin.ConnMetadata, key []byte) (*libplugin.Upstream, error) {
-					return plugin.findAndCreateUpstream(conn, "", key)
-				},
-
-				VerifyHostKeyCallback: func(conn libplugin.ConnMetadata, hostname, netaddr string, key []byte) error {
-					return plugin.verifyHostKey(conn, hostname, netaddr, key)
-				},
-			}, nil
+			skel := libplugin.NewSkelPlugin(plugin.listPipe)
+			return skel.CreateConfig(), nil
 		},
 	})
 }
