+++ plugin/yaml/yaml.go
@@ -8,9 +8,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"regexp"
 
-	"github.com/tg123/sshpiper/libplugin"
 	"gopkg.in/yaml.v3"
 )
 
@@ -73,124 +71,6 @@ func (l *listOrString) UnmarshalYAML(value *yaml.Node) error {
 type yamlPipe struct {
 	From []yamlPipeFrom `yaml:"from,flow"`
 	To   yamlPipeTo     `yaml:"to,flow"`
-}
-
-type skelpipeWrapper struct {
-	plugin *plugin
-
-	pipe *yamlPipe
-}
-
-func (s *skelpipeWrapper) From() []libplugin.SkelPipeFrom {
-	var froms []libplugin.SkelPipeFrom
-	for _, f := range s.pipe.From {
-
-		w := &skelpipePasswordWrapper{
-			plugin: s.plugin,
-			from:   &f,
-			to:     &s.pipe.To,
-		}
-
-		if f.SupportPublicKey() {
-			froms = append(froms, &skelpipePublicKeyWrapper{
-				skelpipePasswordWrapper: *w,
-			})
-		} else {
-			froms = append(froms, w)
-		}
-	}
-	return froms
-}
-
-type skelpipePasswordWrapper struct {
-	plugin *plugin
-	// conn   libplugin.ConnMetadata
-
-	from *yamlPipeFrom
-	to   *yamlPipeTo
-}
-
-type skelpipeToPasswordWrapper struct {
-	plugin *plugin
-
-	username string
-	to       *yamlPipeTo
-}
-
-func (s *skelpipeToPasswordWrapper) User(conn libplugin.ConnMetadata) string {
-	return s.username
-}
-
-func (s *skelpipeToPasswordWrapper) Host(conn libplugin.ConnMetadata) string {
-	return s.to.Host
-}
-
-func (s *skelpipeToPasswordWrapper) IgnoreHostKey(conn libplugin.ConnMetadata) bool {
-	return s.to.IgnoreHostkey
-}
-
-func (s *skelpipeToPasswordWrapper) KnownHosts(conn libplugin.ConnMetadata) ([]byte, error) {
-	return s.plugin.loadFileOrDecodeMany(s.to.KnownHosts, s.to.KnownHostsData, map[string]string{
-		"DOWNSTREAM_USER": conn.User(),
-		"UPSTREAM_USER":   s.username,
-	})
-}
-
-func (s *skelpipePasswordWrapper) Match(conn libplugin.ConnMetadata) (libplugin.SkelPipeTo, error) {
-	user := conn.User()
-
-	matched := s.from.Username == user
-	targetuser := s.to.Username
-
-	if targetuser == "" {
-		targetuser = user
-	}
-
-	if s.from.UsernameRegexMatch {
-		re, err := regexp.Compile(s.from.Username)
-		if err != nil {
-			return nil, err
-		}
-
-		matched = re.MatchString(user)
-
-		if matched {
-			targetuser = re.ReplaceAllString(user, s.to.Username)
-		}
-	}
-
-	if matched {
-		return &skelpipeToPasswordWrapper{
-			plugin:   s.plugin,
-			username: targetuser,
-			to:       s.to,
-		}, nil
-	}
-
-	return nil, nil
-}
-
-type skelpipePublicKeyWrapper struct {
-	skelpipePasswordWrapper
-}
-
-func (s *skelpipePublicKeyWrapper) AuthorizedKeys(conn libplugin.ConnMetadata) ([]byte, error) {
-	return s.plugin.loadFileOrDecodeMany(s.from.AuthorizedKeys, s.from.AuthorizedKeysData, map[string]string{
-		"DOWNSTREAM_USER": conn.User(),
-	})
-}
-
-func (s *skelpipePublicKeyWrapper) TrustedUserCAKeys(conn libplugin.ConnMetadata) ([]byte, error) {
-	return s.plugin.loadFileOrDecodeMany(s.from.TrustedUserCAKeys, s.from.TrustedUserCAKeysData, map[string]string{
-		"DOWNSTREAM_USER": conn.User(),
-	})
-}
-
-func (s *skelpipeToPasswordWrapper) PrivateKey(conn libplugin.ConnMetadata) ([]byte, error) {
-	return s.plugin.loadFileOrDecode(s.to.PrivateKey, s.to.PrivateKeyData, map[string]string{
-		"DOWNSTREAM_USER": conn.User(),
-		"UPSTREAM_USER":   s.username,
-	})
 }
 
 type piperConfig struct {
@@ -304,23 +184,4 @@ func (p *plugin) loadFileOrDecodeMany(files listOrString, base64data listOrStrin
 	}
 
 	return bytes.Join(byteSlices, []byte("\n")), nil
-}
-
-func (p *plugin) listPipe() ([]libplugin.SkelPipe, error) {
-	config, err := p.loadConfig()
-	if err != nil {
-		return nil, err
-	}
-
-	var pipes []libplugin.SkelPipe
-	for _, pipe := range config.Pipes {
-		wrapper := &skelpipeWrapper{
-			plugin: p,
-			pipe:   &pipe,
-		}
-		pipes = append(pipes, wrapper)
-
-	}
-
-	return pipes, nil
 }
